Fixed warnings caused by %u instead of %lu and type casting

+++ lib/libpluto/readwhackmsg.c
@@ -1,4 +1,4 @@
-#include <stdio.h>
+ #include <stdio.h>
 #include <stdlib.h>
 #include "constants.h"
 #include "oswalloc.h"
@@ -37,7 +37,7 @@ void readwhackmsg(char *infile)
 	abuflen = (plen + 3) & ~0x3;
 
 	if(abuflen > sizeof(m1)) {
-	    fprintf(stderr, "whackmsg file has too big a record=%u > %u\n", abuflen, sizeof(m1));
+	  fprintf(stderr, "whackmsg file has too big a record=%u > %lu\n", abuflen, (long unsigned) sizeof(m1));
 	    exit(6);
 	}
